test Extract shared Response mock factory
Cause: `ParseResponse204.test.ts` and `UnsupportedUrl.test.ts` each
defined their own copy of `makeMockResponse` + `IMockArgs` — two
~90-line duplicates that drifted slightly in their JSDoc wording.
CodeRabbit review on commit 2ed8a628 flagged the duplication
(CLAUDE.md `Generic over duplication` rule).

Fix: extract the helper to `_makeMockResponse.ts` (underscore-prefix
follows the project's shared-fixture naming convention —
`_makePhaseFixture`, `_makeBankFixture`). Both test files now
import `makeMockResponse` + `IMockArgs` from the shared module.
Pure refactor — no behavioural change to the mock, just a single
source of truth for the Playwright `Response` stub surface
(`extractRequestMeta` + `parseResponse` read it the same way).

Regression: 394 test suites / 4538 tests GREEN; coverage 97.02 /
95.03 / 97.07 / 98.22 (unchanged); lint + architecture + canaries +
format all clean.
