+++ src/Tests/Unit/Pipeline/Mediator/Network/ParseResponse204.test.ts
@@ -12,99 +12,8 @@
  * `responseBody === null`.
  */
 
-import type { Request, Response } from 'playwright-core';
-
 import { parseResponse } from '../../../../../Scrapers/Pipeline/Mediator/Network/NetworkDiscovery.js';
-
-/** Bundled args for `makeMockResponse` — keeps the helper inside the
- *  3-param ceiling and gives every field its own narrow type. */
-interface IMockArgs {
-  readonly status: number;
-  readonly contentType: string;
-  readonly text: string;
-  readonly url: string;
-  readonly method: 'GET' | 'POST';
-  readonly postData: string;
-}
-
-/**
- * Build a minimal Playwright `Response` stub with the surface
- * `parseResponse` reads: status, headers, url, request().method(),
- * request().postData(), request().headers(), and text().
- *
- * @param args - Bundled mock-response fields.
- * @returns Mock Response cast to the Playwright type.
- */
-function makeMockResponse(args: IMockArgs): Response {
-  /**
-   * Mock request stub — minimal surface used by `extractRequestMeta`.
-   * @returns Request stub with the fields the function reads.
-   */
-  /**
-   * Request-method accessor.
-   * @returns Configured HTTP method.
-   */
-  const requestMethod = (): string => args.method;
-  /**
-   * Request-postData accessor — Playwright returns false for empty.
-   * @returns Configured POST body string, or false when empty.
-   */
-  const requestPostData = (): string | false => {
-    if (args.postData.length > 0) return args.postData;
-    return false;
-  };
-  /**
-   * Request-headers accessor — empty for test isolation.
-   * @returns Empty header map.
-   */
-  const requestHeaders = (): Record<string, string> => ({});
-  /**
-   * Build the Playwright `Request` stub used by `extractRequestMeta`.
-   * @returns Minimal Request stub.
-   */
-  const makeRequest = (): Request => {
-    const stub = {
-      method: requestMethod,
-      postData: requestPostData,
-      headers: requestHeaders,
-    } as unknown as Request;
-    return stub;
-  };
-  /**
-   * Build a header map carrying ONLY the content-type the test
-   * configured. Empty string → no header at all (matches a real 204).
-   * @returns Header map.
-   */
-  const buildHeaders = (): Record<string, string> => {
-    if (args.contentType.length === 0) return {};
-    return { 'content-type': args.contentType };
-  };
-  /**
-   * Response-status accessor.
-   * @returns Configured HTTP status code.
-   */
-  const responseStatus = (): number => args.status;
-  /**
-   * Response-URL accessor.
-   * @returns Configured URL.
-   */
-  const responseUrl = (): string => args.url;
-  /**
-   * Response-text accessor — Playwright resolves async; we return a
-   * pre-resolved promise so the lint `require-await` rule passes
-   * without an artificial await.
-   * @returns Configured body text.
-   */
-  const responseText = (): Promise<string> => Promise.resolve(args.text);
-  const stub = {
-    status: responseStatus,
-    headers: buildHeaders,
-    url: responseUrl,
-    request: makeRequest,
-    text: responseText,
-  } as unknown as Response;
-  return stub;
-}
+import { type IMockArgs, makeMockResponse } from './_makeMockResponse.js';
 
 /** Synthetic args helper — sets the live-Hapoalim shape defaults. */
 const DEFAULT_ARGS: IMockArgs = {
